backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.0.0
python-dateutil==2.8.2
Werkzeug==2.3.7